role based crud, crud with post req only

+++ app_admin/views.py
@@ -10,7 +10,9 @@ def profile(request):
     if request.user.is_authenticated:
         userscreens = request.session['userscreens']
         modules = request.session['modules']
-        return render(request, 'adminlte/profile.html', {'modules': modules, 'userscreens': userscreens})
+        role = request.user.user_role_id.id
+        print(role)
+        return render(request, 'adminlte/profile.html', {'modules': modules, 'userscreens': userscreens, 'role': role})
     return redirect('login')
 
 # product
@@ -21,6 +23,9 @@ def product(request):
         products = Product.objects.all()
         categories = Category.objects.all()
         manufacturers = Manufacturer.objects.all()
+        
+        role = request.user.user_role_id.id
+
         if request.method == 'POST':
             name = request.POST['name']
             category = request.POST['category']
@@ -35,29 +40,33 @@ def product(request):
                 product.save()
             return redirect('products')
 
-        return render(request, 'adminlte/product.html', {'modules': modules,'products': products, 'categories': categories, 'manufacturers': manufacturers, 'userscreens': userscreens})
+        return render(request, 'adminlte/product.html', {'modules': modules,'products': products, 'categories': categories, 'manufacturers': manufacturers, 'userscreens': userscreens, 'sideScreen': 'All Products', 'sideSub': 'Product', 'sideModule': 'Products', 'role': role})
     return redirect('login')
 
 
 # edit product
-def edit_product(request, id):
+def edit_product(request):
     if request.user.is_authenticated:
-        product = Product.objects.get(id=id)
-        product.name = request.POST['name']
-        product.category_id = request.POST['category']
-        product.manufacturer_id = request.POST['manufacturer']
-        product.unit = request.POST['unit']
-        product.price = request.POST['price']
-        product.save()
+        if request.method == 'POST':
+            id = request.POST['id']
+            product = Product.objects.get(id=id)
+            product.name = request.POST['name']
+            product.category_id = request.POST['category']
+            product.manufacturer_id = request.POST['manufacturer']
+            product.unit = request.POST['unit']
+            product.price = request.POST['price']
+            product.save()
         return redirect('products')
     return redirect('login')
 
 
 # delete product
-def delete_product(request, id):
+def delete_product(request):
     if request.user.is_authenticated:
-        product = Product.objects.get(id=id)
-        product.delete()
+        if request.method == 'POST':
+            id = request.POST['id']
+            product = Product.objects.get(id=id) 
+            product.delete()
         return redirect('products')
     return redirect('login')
 
@@ -68,6 +77,9 @@ def category(request):
         userscreens = request.session['userscreens']
         modules = request.session['modules']
         categories = Category.objects.all()
+        
+        role = request.user.user_role_id.id
+
         if request.method == 'POST':
             name = request.POST['name']
             if Category.objects.filter(name=name).exists():
@@ -76,23 +88,26 @@ def category(request):
                 category = Category(name=name)
                 category.save()
             return redirect('categories')
-        return render(request, 'adminlte/category.html',{'categories': categories, 'modules':modules, 'userscreens':userscreens})
+        return render(request, 'adminlte/category.html',{'categories': categories, 'modules':modules, 'userscreens':userscreens , 'sideScreen': 'Category', 'role': role})
     return redirect('login')
 
 
 # edit category
-def edit_category(request, id):
+def edit_category(request):
     if request.user.is_authenticated:
+        id = request.POST['id']
         category = Category.objects.get(id=id)
         category.name = request.POST['name']
         category.save()
+        messages.info(request, 'Category updated!')
         return redirect('categories')
     return redirect('login')
 
 
 # delete category
-def delete_category(request, id):
+def delete_category(request):
     if request.user.is_authenticated:
+        id = request.POST['id']
         category = Category.objects.get(id=id)
         category.delete()
         return redirect('categories')
@@ -105,6 +120,9 @@ def manufacturer(request):
         userscreens = request.session['userscreens']
         modules = request.session['modules']
         manufacturers = Manufacturer.objects.all()
+
+        role = request.user.user_role_id.id
+
         if request.method == 'POST':
             name = request.POST['name']
             if Manufacturer.objects.filter(name=name).exists():
@@ -113,24 +131,28 @@ def manufacturer(request):
                 manufacturer = Manufacturer(name=name)
                 manufacturer.save()
             return redirect('manufacturers')
-        return render(request, 'adminlte/manufacturer.html',{'manufacturers': manufacturers, 'modules':modules, 'userscreens':userscreens})
+        return render(request, 'adminlte/manufacturer.html',{'manufacturers': manufacturers, 'modules':modules, 'userscreens':userscreens, 'sideScreen': 'All Manufacturers', 'sideSub': 'Manufacturer', 'sideModule': 'Products', 'role': role})
     return redirect('login')
 
 
 # edit manufacturer
-def edit_manufacturer(request, id):
+def edit_manufacturer(request):
     if request.user.is_authenticated:
-        manufacturer = Manufacturer.objects.get(id=id)
-        manufacturer.name = request.POST['name']
-        manufacturer.save()
+        if request.method == 'POST':
+            id = request.POST['id']
+            manufacturer = Manufacturer.objects.get(id=id)
+            manufacturer.name = request.POST['name']
+            manufacturer.save()
         return redirect('manufacturers')
     return redirect('login')
 
 
 # delete manufacturer
-def delete_manufacturer(request, id):
+def delete_manufacturer(request):
     if request.user.is_authenticated:
-        manufacturer = Manufacturer.objects.get(id=id)
-        manufacturer.delete()
+        if request.method == 'POST':
+            id = request.POST['id']
+            manufacturer = Manufacturer.objects.get(id=id)
+            manufacturer.delete()
         return redirect('manufacturers')
     return redirect('login')
